Replication of racetrack working
track generation now triggered in tick method instead of begin play to ensure racetrack actor has replicated to the client before attempting to spawn the track. This could possible be changed to use RepUsing instead. Need to create a "HasGenerated" function so it is clear if the track has been generated or not yet.